changes in controllers and service

+++ src/main/java/com/diplomski/bookingkidsparty/app/service/PaymentService.java
@@ -4,7 +4,7 @@ import com.stripe.exception.StripeException;
 
 public interface PaymentService {
 
-	String createPaymentIntent(Long amount, String username, String email) throws StripeException;
+	String createPaymentIntent(double amount, String username, String email) throws StripeException;
 
 	void refunde(String clientSecret) throws StripeException;
